JUST SHUT UP FOR GOODNESS SAKE

+++ src/public/tier0/threadtools.h
@@ -402,12 +402,12 @@ PLATFORM_INTERFACE int32 ThreadInterlockedExchangeAdd( int32 volatile *, int32 v
 PLATFORM_INTERFACE int32 ThreadInterlockedCompareExchange( int32 volatile *, int32 value, int32 comperand ) NOINLINE;
 PLATFORM_INTERFACE bool ThreadInterlockedAssignIf( int32 volatile *, int32 value, int32 comperand ) NOINLINE;
 #else
-PLATFORM_INTERFACE int64_t ThreadInterlockedIncrement( int64_t volatile * ) NOINLINE;
-PLATFORM_INTERFACE int64_t ThreadInterlockedDecrement( int64_t volatile * ) NOINLINE;
-PLATFORM_INTERFACE int64_t ThreadInterlockedExchange( int64 volatile *, int64_t value ) NOINLINE;
-PLATFORM_INTERFACE int64_t ThreadInterlockedExchangeAdd( int64_t volatile *, int64_t value ) NOINLINE;
-PLATFORM_INTERFACE int64_t ThreadInterlockedCompareExchange( int64_t volatile *, int64_t value, int64_t comperand ) NOINLINE;
-PLATFORM_INTERFACE bool ThreadInterlockedAssignIf( int64_t volatile *, int64_t value, int64_t comperand ) NOINLINE;
+PLATFORM_INTERFACE int64_t ThreadInterlockedIncrement( int volatile * ) NOINLINE;
+PLATFORM_INTERFACE int64_t ThreadInterlockedDecrement( int volatile * ) NOINLINE;
+PLATFORM_INTERFACE int64_t ThreadInterlockedExchange( int volatile *, int64_t value ) NOINLINE;
+PLATFORM_INTERFACE int64_t ThreadInterlockedExchangeAdd( int volatile *, int64_t value ) NOINLINE;
+PLATFORM_INTERFACE int64_t ThreadInterlockedCompareExchange( int volatile *, int64_t value, int64_t comperand ) NOINLINE;
+PLATFORM_INTERFACE bool ThreadInterlockedAssignIf( int volatile *, int64_t value, int64_t comperand ) NOINLINE;
 #endif
 #endif
